# doc/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Swagger Movies",
    "description": "Please use account: `admin`, password: `123@qweQ!` to login and get token. Please config environment in the file `.env` before migrate or seed db.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "tin.quachtrung@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:{port}/api",
      "variables": {
        "port": {
          "enum": [
            "6969",
            "8082"
          ],
          "default": "8082"
        }
      }
    }
  ],
...
  "paths": {
...
    "/manageshowtime": {
      "get": {
        "tags": [
          "Mangement Showtime"
        ],
        "summary": "Get information showtime",
        "parameters": [
          {
            "name": "showtimeCode",
            "in": "query",
            "description": "Showtime Code",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/manageshowtime/getshowtimebycineplex": {
      "get": {
        "tags": [
          "Mangement Showtime"
        ],
        "summary": "Get information showtime by cineplex",
        "parameters": [
          {
            "name": "cinplexCode",
            "in": "query",
            "description": "Cineplex Code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupCode",
            "in": "query",
            "description": "group Code",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/manageshowtime/getshowtimebyFilm": {
      "get": {
        "tags": [
          "Mangement Showtime"
        ],
        "summary": "Get information showtime by film",
        "parameters": [
          {
            "name": "movieCode",
            "in": "query",
            "description": "Movie Code",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/manageshowtime/create": {
      "post": {
        "tags": [
          "Mangement Showtime"
        ],
        "summary": "Create Showtime",
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "requestBody": {
          "description": "Create Showtime",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Showtime"
              }
            }
          },
          "required": true,
          "x-codegen-request-body-name": "body"
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/managemovie": {
      "get": {
        "tags": [
          "Mangement Movie"
        ],
        "summary": "Get list Movie",
        "parameters": [
          {
            "name": "movieName",
            "in": "query",
            "description": "Movie Name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupCode",
            "in": "query",
            "description": "group Code",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/managemovie/paging": {
      "get": {
        "tags": [
          "Mangement Movie"
        ],
        "summary": "Get list Movie",
        "parameters": [
          {
            "name": "movieName",
            "in": "query",
            "description": "Movie Name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupCode",
            "in": "query",
            "description": "group Code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page of list user",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of list user",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/managemovie/date": {
      "get": {
        "tags": [
          "Mangement Movie"
        ],
        "summary": "Get list Movie",
        "parameters": [
          {
            "name": "movieName",
            "in": "query",
            "description": "Movie Name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupCode",
            "in": "query",
            "description": "group Code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page of list user",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of list user",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "description": "The date begin",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "description": "The date end",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/managemovie/delete": {
      "delete": {
        "tags": [
          "Mangement Movie"
        ],
        "summary": "Delete Movie",
        "parameters": [
          {
            "name": "movieCode",
            "in": "query",
            "description": "The Movie Code",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/managemovie/create": {
      "post": {
        "tags": [
          "Mangement Movie"
        ],
        "summary": "Delete Movie",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/managemovie/update": {
      "post": {
        "tags": [
          "Mangement Movie"
        ],
        "summary": "update Movie",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Showtime": {
        "type": "object",
        "properties": {
          "startTime": {
            "type": "string",
            "format": "date"
          },
          "theaterCode": {
            "type": "string"
          },
          "movieCode": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        }
      },
...
      "Movie": {
        "type": "object",
        "properties": {
          "movieCode": {
            "type": "number"
          },
          "movieName": {
            "type": "string"
          },
          "movieAlias": {
            "type": "string"
          },
          "trailer": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "groupCode": {
            "type": "string"
          },
          "openingDay": {
            "type": "string",
            "format": "date"
          },
          "duration": {
            "type": "number"
          },
          "rate": {
            "type": "number"
          },
          "imageUrl": {
            "type": "string",
            "format": "base64"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}